+++ src/axum_routes/routes/logs_routes/browse_logs_paginated/browse_logs_paginated_sql.rs
@@ -3,7 +3,7 @@ use mysql::prelude::Queryable;
 use tokio::sync::MutexGuard;
 use crate::axum_routes::routes::logs_routes::browse_logs_paginated::browse_logs_paginated_structs::SingleLogObject;
 
-pub fn browse_logs_paginated_sql(page_number : u16, rows_per_page : u16, pool : &mut MutexGuard<PooledConn>) -> mysql::Result<Vec<SingleLogObject>, Error> {
+pub fn browse_logs_paginated_sql(page_number : u32, rows_per_page : u32, pool : &mut MutexGuard<PooledConn>) -> mysql::Result<Vec<SingleLogObject>, Error> {
     match pool.query_map(format_the_query(page_number, rows_per_page),
         |(id, contents, date_time)| {
             SingleLogObject {
@@ -20,6 +20,6 @@ pub fn browse_logs_paginated_sql(page_number : u16, rows_per_page : u16, pool : 
     }
 }
 
-fn format_the_query(page_number : u16, rows_per_page : u16) -> String {
+fn format_the_query(page_number : u32, rows_per_page : u32) -> String {
     return format!("SELECT id, contents, date_time FROM `ugo_logs` ORDER BY id DESC LIMIT {} OFFSET {}", rows_per_page, rows_per_page * page_number)
 }
