+++ src/axum_routes/routes/orders_routes/get_filtered_orders_by_page/get_filtered_orders_by_page_sql.rs
@@ -4,13 +4,13 @@ use tokio::sync::MutexGuard;
 use crate::structs::structs::{BasicPartGetAll, FormattedObject};
 use crate::structs::tool_functions::collect_group_notes;
 
-fn format_the_query(page_number : u16, rows_per_page : u16, filter_type : &String, filter_query : &String) -> String {
+fn format_the_query(page_number : u32, rows_per_page : u32, filter_type : &String, filter_query : &String) -> String {
     if filter_type == "date_time_added" {
         return format!(r#"SELECT id, request_status, customer_name, customer_email, customer_self_description, date_time_added FROM ugo_customers_request WHERE DATE(`date_time_added`) = DATE("{}") ORDER BY id DESC LIMIT {} OFFSET {}"#, filter_query, rows_per_page, rows_per_page * page_number)
     }
     return format!("SELECT id, request_status, customer_name, customer_email, customer_self_description, date_time_added FROM `ugo_customers_request` WHERE {} LIKE '%{}%' ORDER BY id DESC LIMIT {} OFFSET {}", filter_type, filter_query, rows_per_page, rows_per_page * page_number)
 }
-pub fn get_filtered_orders_by_page_sql(page_number : u16, rows_per_page : u16, filter_type : &String, filter_query : &String, pool : &mut MutexGuard<PooledConn>) -> mysql::Result<Vec<FormattedObject>, Error> {
+pub fn get_filtered_orders_by_page_sql(page_number : u32, rows_per_page : u32, filter_type : &String, filter_query : &String, pool : &mut MutexGuard<PooledConn>) -> mysql::Result<Vec<FormattedObject>, Error> {
     match pool.query_map(format_the_query(page_number, rows_per_page, filter_type, filter_query),
                          |(id, request_status, customer_name, customer_email, customer_self_description, date_time_added)| {
                              BasicPartGetAll {
